+++ koreantime/app/page.tsx
@@ -1,7 +1,12 @@
 import getCurrentUser from "./actions/getCurrentUser";
-import LoginForm from "./login/page";
+import LoginForm from "./login/LoginForm";
 
 export default async function Home() {
     const currentUser = await getCurrentUser();
-    return <LoginForm currentUser={currentUser} />;
+
+    return (
+        <>
+            <LoginForm currentUser={currentUser} />
+        </>
+    );
 }
